+++ inlang/source-code/doc-layout-component/src/index.ts
@@ -1,0 +1,1 @@
+export { default as InlangDocLayout } from "./stories/inlang-doc-layout.js"