+++ inlang/source-code/editor/CHANGELOG.md
@@ -4,6 +4,25 @@
 
 ### Patch Changes
 
+- Updated dependencies [9165e64]
+  - @lix-js/client@1.4.0
+  - @inlang/settings-component@1.0.5
+
+## null
+
+### Patch Changes
+
+- Updated dependencies [c92bde5]
+- Updated dependencies [7bd98e4]
+- Updated dependencies [2508271]
+- Updated dependencies [0b4af82]
+  - @inlang/settings-component@1.0.3
+  - @lix-js/client@1.3.0
+
+## null
+
+### Patch Changes
+
 - Updated dependencies [4941ca4]
   - @lix-js/client@1.2.1