+++ inlang/source-code/plugins/json/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @inlang/plugin-json
 
+## 5.1.31
+
+### Patch Changes
+
+- Updated dependencies [3c959bc]
+  - @inlang/sdk@0.34.2
+
+## 5.1.30
+
+### Patch Changes
+
+- @inlang/sdk@0.34.1
+
+## 5.1.29
+
+### Patch Changes
+
+- Updated dependencies [5b8c053]
+  - @inlang/sdk@0.34.0
+
+## 5.1.28
+
+### Patch Changes
+
+- cd9a3e1: add internal links
+  - @inlang/sdk@0.33.1
+
 ## 5.1.27
 
 ### Patch Changes