+++ inlang/source-code/website/CHANGELOG.md
@@ -4,6 +4,20 @@
 
 ### Patch Changes
 
+- Updated dependencies [9165e64]
+  - @lix-js/client@1.4.0
+
+## null
+
+### Patch Changes
+
+- Updated dependencies [7bd98e4]
+  - @lix-js/client@1.3.0
+
+## null
+
+### Patch Changes
+
 - Updated dependencies [4941ca4]
   - @lix-js/client@1.2.1